% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHexGridVariant.r
\name{sHexGridVariant}
\alias{sHexGridVariant}
\title{Function to define a variant of a supra-hexagonal grid}
\usage{
sHexGridVariant(r = NULL, nHex = NULL, shape = c("suprahex",
"triangle",
"diamond", "hourglass", "trefoil", "ladder", "butterfly"))
}
\arguments{
\item{r}{an integer specifying the radius in a supra-hexagonal grid}

\item{nHex}{the number of input hexagons in the grid}

\item{shape}{the grid shape, either "suprahex" for the suprahex itself,
or its variants (including "triangle" for the triangle-shaped variant,
"diamond" for the diamond-shaped variant, "hourglass" for the
hourglass-shaped variant, "trefoil" for the trefoil-shaped variant,
"ladder" for the ladder-shaped variant, and "butterfly" for the
butterfly-shaped variant)}
}
\value{
an object of class "sHex", a list with following components:
\itemize{
\item{\code{r}: the grid radius}
\item{\code{nHex}: the total number of hexagons in the grid. It may
differ from the input value; actually it is always no less than the
input one to ensure a supra-hexagonal grid exactly formed}
\item{\code{centroid}: the 2D coordinates of the grid centroid}
\item{\code{stepCentroid}: a vector with the length of nHex. It stores
how many steps a hexagon is awawy from the grid centroid ('1' for the
centroid itself). Starting with the centroid, it orders outward. Also,
for those hexagons of the same step, it orders from the rightmost in an
anti-clock wise}
\item{\code{angleCentroid}: a vector with the length of nHex. It stores
the angle a hexagon is in terms of the grid centroid ('0' for the
centroid itself). For those hexagons of the same step, it orders from
the rightmost in an anti-clock wise}
\item{\code{coord}: a matrix of nHex x 2 with each row specifying the
2D coordinates of a hexagon in the grid. The order of rows is the same
as 'centroid' above}
\item{\code{call}: the call that produced this result}
}
}
\description{
\code{sHexGridVariant} is supposed to define a variant of a
supra-hexagonal map grid. In essence, it is the subset of the
supra-hexagon.
}
\note{
none
}
\examples{
\dontrun{
library(ggplot2)

#geom_polygon(color="black", fill=NA)

# For "supraHex" shape itself
sHex <- sHexGridVariant(r=8, shape="suprahex")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_suprahex <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "trangle" shape
sHex <- sHexGridVariant(r=8, shape="triangle")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_trangle <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "diamond" shape
sHex <- sHexGridVariant(r=8, shape="diamond")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_diamond <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "hourglass" shape
sHex <- sHexGridVariant(r=8, shape="hourglass")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_hourglass <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "trefoil" shape
sHex <- sHexGridVariant(r=8, shape="trefoil")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_trefoil <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "ladder" shape
sHex <- sHexGridVariant(r=8, shape="ladder")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_ladder <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# For "butterfly" shape
sHex <- sHexGridVariant(r=8, shape="butterfly")
df_polygon <- sHexPolygon(sHex)
df_coord <- data.frame(sHex$coord, index=1:nrow(sHex$coord))
gp_butterfly <- ggplot(data=df_polygon, aes(x,y,group=index)) +
geom_polygon(aes(fill=factor(stepCentroid\%\%2))) +
coord_fixed(ratio=1) + theme_void() + theme(legend.position="none") +
geom_text(data=df_coord, aes(x,y,label=index), color="white")

# combined visuals
library(gridExtra)
grid.arrange(grobs=list(gp_suprahex,gp_trangle,gp_diamond,gp_hourglass,gp_trefoil,gp_ladder,gp_butterfly),
layout_matrix=rbind(c(1,1,3,4),c(1,1,3,4),c(5,5,7,7),c(5,5,7,7),c(6,6,2,NA)),
nrow=5, ncol=4)
}
}
\seealso{
\code{\link{sHexGrid}}
}

