% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHexPolygon.r
\name{sHexPolygon}
\alias{sHexPolygon}
\title{Function to extract polygon location per hexagon within a supra-hexagonal grid}
\usage{
sHexPolygon(sObj, area.size = 1)
}
\arguments{
\item{sObj}{an object of class "sMap" or "sInit" or "sTopol" or "sHex"}

\item{area.size}{an integer or a vector specifying the area size of
each hexagon}
}
\value{
a tibble of 7 columns
('index','x','y','node','edge','stepCentroid','angleCentroid') storing
polygon location per hexagon. 'node' for nodes (including
n1,n2,n3,n4,n5,n6), and 'edge' for a list-column where each is a tibble
with a single column 'edge' containing two rows (such as edges 'e12'
and 'e16' for the node 'n1').
}
\description{
\code{sHexPolygon} is supposed to extract polygon location per hexagon
within a supra-hexagonal grid
}
\note{
None
}
\examples{
sObj <- sTopology(xdim=4, ydim=4, lattice="hexa", shape="suprahex")
df_polygon <- sHexPolygon(sObj, area.size=1)
}
\seealso{
\code{\link{sHexGridVariant}}, \code{\link{sPipeline}}
}
