% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visHexBarplot.r
\name{visHexBarplot}
\alias{visHexBarplot}
\title{Function to visualise codebook matrix using barplot for all hexagons or a specific one}
\usage{
visHexBarplot(
sObj,
which.hexagon = NULL,
which.hexagon.highlight = NULL,
height = 7,
margin = rep(0.1, 4),
colormap = c("customized", "bwr", "jet", "gbr", "wyr", "br", "yr",
"rainbow", "wb"),
customized.color = "red",
zeropattern.color = "gray",
gp = grid::gpar(cex = 0.7, font = 1, col = "black"),
bar.text.cex = 0.8,
bar.text.srt = 90,
newpage = TRUE
)
}
\arguments{
\item{sObj}{an object of class "sMap" or "sTopol" or "sInit"}

\item{which.hexagon}{the integer specifying which hexagon to display.
If NULL, all hexagons will be visualised}

\item{which.hexagon.highlight}{an integer vector specifying which
hexagons are labelled. If NULL, all hexagons will be labelled}

\item{height}{a numeric value specifying the height of device}

\item{margin}{margins as units of length 4 or 1}

\item{colormap}{short name for the predifined colormap, and
"customized" for custom input (see the next 'customized.color'). The
predifined colormap can be one of "jet" (jet colormap), "bwr"
(blue-white-red colormap), "gbr" (green-black-red colormap), "wyr"
(white-yellow-red colormap), "br" (black-red colormap), "yr"
(yellow-red colormap), "wb" (white-black colormap), and "rainbow"
(rainbow colormap, that is, red-yellow-green-cyan-blue-magenta).
Alternatively, any hyphen-separated HTML color names, e.g.
"blue-black-yellow", "royalblue-white-sandybrown",
"darkgreen-white-darkviolet". A list of standard color names can be
found in \url{https://html-color-codes.info/color-names/index.html}}

\item{customized.color}{the customized color for pattern visualisation}

\item{zeropattern.color}{the color for zero horizental line}

\item{gp}{an object of class "gpar". It is the output from a call to
the function "gpar" (i.e., a list of graphical parameter settings)}

\item{bar.text.cex}{a numerical value giving the amount by which bar
text should be magnified relative to the default (i.e., 1)}

\item{bar.text.srt}{a numerical value giving the angle by which bar
text should be orientated}

\item{newpage}{logical to indicate whether to open a new page. By
default, it sets to true for opening a new page}
}
\value{
invisible
}
\description{
\code{visHexBarplot} is supposed to visualise codebook matrix using
barplot for all hexagons or a specific one
}
\note{
none
}
\examples{
# 1) generate data with an iid matrix of 1000 x 9
data <- cbind(matrix(rnorm(1000*3,mean=0,sd=1), nrow=1000, ncol=3),
matrix(rnorm(1000*3,mean=0.5,sd=1), nrow=1000, ncol=3),
matrix(rnorm(1000*3,mean=-0.5,sd=1), nrow=1000, ncol=3))
colnames(data) <- c("S1","S1","S1","S2","S2","S2","S3","S3","S3")

# 2) sMap resulted from using by default setup
sMap <- sPipeline(data=data)

# 3) plot codebook patterns using different types
# 3a) for all hexagons
visHexBarplot(sMap)
# 3b) only for the first hexagon
visHexBarplot(sMap, which.hexagon=1)
}
\seealso{
\code{\link{sPipeline}}, \code{\link{visColormap}}
}
