% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTreeBSclust.r
\name{visTreeBSclust}
\alias{visTreeBSclust}
\title{Function to obtain clusters from a bootstrapped tree}
\usage{
visTreeBSclust(
tree_bs,
bootstrap.cutoff = 80,
max.fraction = 1,
min.size = 3,
visTree = TRUE,
plot.phylo.arg = NULL,
nodelabels.arg = NULL,
verbose = TRUE,
...
)
}
\arguments{
\item{tree_bs}{an "phylo" object storing a bootstrapped tree}

\item{bootstrap.cutoff}{an integer specifying bootstrap-derived
clusters}

\item{max.fraction}{the maximum fraction of leaves contained in a
cluster}

\item{min.size}{the minumum number of leaves contained in a cluster}

\item{visTree}{logical to indicate whether the tree will be visualised.
By default, it sets to true for display}

\item{plot.phylo.arg}{a list of main parameters used in the function
"ape::plot.phylo" \url{http://rdrr.io/cran/ape/man/plot.phylo.html}.
See 'Note' below for details on the parameters}

\item{nodelabels.arg}{a list of main parameters used in the function
"ape::nodelabels" \url{http://rdrr.io/cran/ape/man/nodelabels.html}.
See 'Note' below for details on the parameters}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{...}{additional "ape::plot.phylo" parameters}
}
\value{
a data frame following components:
\itemize{
\item{\code{Samples}: the labels for tip nodes (samples)}
\item{\code{Clusters}: the clusters each tip node belongs to;
unassigned tip nodes will be the cluster called 'C0'}
\item{\code{Clans}: the internal node id for each cluster}
}
}
\description{
\code{visTreeBSclust} is supposed to obtain clusters from a
bootstrapped tree.
}
\note{
A list of main parameters used in the function "ape::plot.phylo":
\itemize{
\item{"type": a character string specifying the type of phylogeny to be
drawn; it must be one of "phylogram" (the default), "cladogram", "fan",
"unrooted", "radial" or any unambiguous abbreviation of these}
\item{"direction": a character string specifying the direction of the
tree. Four values are possible: "rightwards" (the default),
"leftwards", "upwards", and "downwards"}
\item{"lab4ut": (= labels for unrooted trees) a character string
specifying the display of tip labels for unrooted trees: either
"horizontal" where all labels are horizontal (the default), or "axial"
where the labels are displayed in the axis of the corresponding
terminal branches. This option has an effect only if type = "unrooted"}
\item{"edge.color": a vector of mode character giving the colours used
to draw the branches of the plotted phylogeny. These are taken to be in
the same order than the component edge of phy. If fewer colours are
given than the length of edge, then the colours are recycled}
\item{"edge.width": a numeric vector giving the width of the branches
of the plotted phylogeny. These are taken to be in the same order than
the component edge of phy. If fewer widths are given than the length of
edge, then these are recycled}
\item{"edge.lty": same than the previous argument but for line types;
1: plain, 2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash}
\item{"font": an integer specifying the type of font for the labels: 1
(plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic)}
\item{"cex": a numeric value giving the factor scaling of the tip and
node labels (Character EXpansion). The default is to take the current
value from the graphical parameters}
\item{"adj": a numeric specifying the justification of the text strings
of the labels: 0 (left-justification), 0.5 (centering), or 1
(right-justification). This option has no effect if type="unrooted". If
NULL (the default) the value is set with respect of direction (see
details)}
\item{"srt": a numeric giving how much the labels are rotated in
degrees (negative values are allowed resulting in clock-like rotation);
the value has an effect respectively to the value of direction (see
Examples). This option has no effect if type="unrooted"}
\item{"no.margin": a logical. If TRUE, the margins are set to zero and
the plot uses all the space of the device}
\item{"label.offset": a numeric giving the space between the nodes and
the tips of the phylogeny and their corresponding labels. This option
has no effect if type="unrooted"}
\item{"rotate.tree": for "fan", "unrooted", or "radial" trees: the
rotation of the whole tree in degrees (negative values are accepted}
}
A list of main parameters used in the function "ape::nodelabels":
\itemize{
\item{"text": a vector of mode character giving the text to be printed.
By default, the labels for internal nodes (see "node.label"), that is,
the bootstrap values associated with internal nodes}
\item{"node": a vector of mode numeric giving the numbers of the nodes
where the text or the symbols are to be printed. By default, indexes
for internal nodes, that is, (\eqn{Ntip}+1):(\eqn{Ntip}+\eqn{Nnode}),
where \eqn{Ntip} is the number of tip nodes and \eqn{Nnode} for the
number of internal nodes}
\item{"adj": one or two numeric values specifying the horizontal and
vertical, respectively, justification of the text or symbols. By
default, the text is centered horizontally and vertically. If a single
value is given, this alters only the horizontal position of the text}
\item{"frame": a character string specifying the kind of frame to be
printed around the text. This must be one of "rect" (the default),
"circle", "none", or any unambiguous abbreviation of these}
\item{"cex": a numeric value giving the factor scaling of the tip and
node labels (Character EXpansion). The default is to take the current
value from the graphical parameters}
\item{"font": an integer specifying the type of font for the labels: 1
(plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic)}
\item{"col": a character string giving the color to be used for the
text or the plotting symbols; this is eventually recycled}
\item{"bg": a character string giving the color to be used for the
background of the text frames or of the plotting symbols if it applies;
this is eventually recycled. It can be one of "jet" (jet colormap),
"bwr" (blue-white-red colormap), "gbr" (green-black-red colormap),
"wyr" (white-yellow-red colormap), "br" (black-red colormap), "yr"
(yellow-red colormap), "wb" (white-black colormap), and "rainbow"
(rainbow colormap, that is, red-yellow-green-cyan-blue-magenta).
Alternatively, any hyphen-separated HTML color names, e.g.
"blue-black-yellow", "royalblue-white-sandybrown",
"darkgreen-white-darkviolet". A list of standard color names can be
found in \url{https://html-color-codes.info/color-names/index.html}}
}
}
\examples{

# 1) generate an iid normal random matrix of 100x10 
data <- matrix( rnorm(100*10,mean=0,sd=1), nrow=100, ncol=10)
colnames(data) <- paste(rep('S',10), seq(1:10), sep="")
data <- t(data)


\dontrun{
# 2) build neighbor-joining tree with bootstrap values and visualise it by default
tree_bs <- visTreeBootstrap(data)

# 3) obtain clusters from a bootstrapped tree
res <- visTreeBSclust(tree_bs, bootstrap.cutoff=80)
## hide tip labels and modify the font of internal node labels
res <- visTreeBSclust(tree_bs, bootstrap.cutoff=80,
nodelabels.arg=list(cex=0.4), show.tip.label=FALSE)
}
}
\seealso{
\code{\link{visTreeBootstrap}}
}
