% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visHexGrid.r
\name{visHexGrid}
\alias{visHexGrid}
\title{Function to visualise a supra-hexagonal grid}
\usage{
visHexGrid(
hbin,
area.size = 1,
border.color = NULL,
fill.color = NULL,
lty = 1,
lwd = 1,
lineend = "round",
linejoin = "round"
)
}
\arguments{
\item{hbin}{an object of class "hexbin"}

\item{area.size}{an inteter or a vector specifying the area size of
each hexagon}

\item{border.color}{the border color for each hexagon}

\item{fill.color}{the filled color for each hexagon}

\item{lty}{the line type for each hexagon. 0 for 'blank', 1 for
'solid', 2 for 'dashed', 3 for 'dotted', 4 for 'dotdash', 5 for
'longdash', 6 for 'twodash'}

\item{lwd}{the line width for each hexagon}

\item{lineend}{the line end style for each hexagon. It can be one of
'round', 'butt' and 'square'}

\item{linejoin}{the line join style for each hexagon. It can be one of
'round', 'mitre' and 'bevel'}
}
\value{
invisible
}
\description{
\code{visHexGrid} is supposed to visualise a supra-hexagonal grid
}
\note{
none
}
\examples{
# 1) generate an iid normal random matrix of 100x10 
data <- matrix( rnorm(100*10,mean=0,sd=1), nrow=100, ncol=10)
colnames(data) <- paste(rep('S',10), seq(1:10), sep="")

# 2) sMap resulted from using by default setup
sMap <- sPipeline(data=data)

# 3) create an object of "hexbin" class from sMap
dat <- data.frame(sMap$coord)
xdim <- sMap$xdim
ydim <- sMap$ydim
hbin <- hexbin::hexbin(dat$x, dat$y, xbins=xdim-1,
shape=sqrt(0.75)*ydim/xdim)

# 4) visualise hbin object
vp <- hexbin::hexViewport(hbin)
visHexGrid(hbin)
}
\seealso{
\code{\link{visHexComp}}
}
